# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /saved-credentials:
    get:
      operationId: listSavedCredentials
      summary: List saved credentials (without secrets)
      tags: [saved-credentials]
      responses:
        "200":
          description: List of saved credentials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedCredential"
    post:
      operationId: createSavedCredential
      summary: Create a saved credential
      tags: [saved-credentials]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavedCredentialBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedCredential"
  /saved-credentials/{id}:
    put:
      operationId: updateSavedCredential
      summary: Update a saved credential
      tags: [saved-credentials]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSavedCredentialBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedCredential"
    delete:
      operationId: deleteSavedCredential
      summary: Delete a saved credential
      tags: [saved-credentials]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...